refactor: moved editor components to the editor folder
